% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childSurv-package.R
\docType{package}
\name{childSurv-package}
\alias{childSurv}
\alias{childSurv-package}
\title{childSurv: Fit survival models for child mortality}
\description{
Fit survival models for child mortality using vital registration, household surveys, and other available data types.
}
\author{
\strong{Maintainer}: Katie Paulson \email{krpaul@uw.edu} (\href{https://orcid.org/0000-0002-6521-7994}{ORCID})

}
\keyword{internal}
