% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childSurvLL-package.R
\docType{package}
\name{childSurvLL-package}
\alias{childSurvLL}
\alias{childSurvLL-package}
\title{childSurvLL: Fit log-logistic survival models for child mortality}
\description{
Fit log-logistic survival models for child mortality using vital registration, household surveys, and other available data types.
}
\author{
\strong{Maintainer}: Katie Paulson \email{krpaul@uw.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
