% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_adjustment.R
\name{mm_adjustment}
\alias{mm_adjustment}
\title{Apply missing mothers adjustment}
\usage{
mm_adjustment(fit_pssst, ratios, survival_model, sy)
}
\arguments{
\item{fit_pssst}{fit object from the output of pssst::surv_synthetic()}

\item{ratios}{data.frame with columns 'year' and 'ratio' indicating the
ratios to apply, equal to adjusted to unadjusted U5MR.}

\item{survival_model}{either 'piecewise_exponential' or 'log_logistic'}

\item{sy}{survey year}
}
\value{
modified fit_pssst with new U5MR, IMR, NMR, and survival parameters.
}
\description{
Apply missing mothers adjustment to pssst output for piecewise-exponential
or log-logistic survival models.
}
