% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_logistic.R
\name{log_logistic}
\alias{log_logistic}
\title{Fit log-logistic model}
\usage{
log_logistic(data_vr, time_model, start_year = 1950, end_year = 2030)
}
\arguments{
\item{data_vr}{input data}

\item{time_model}{spline or rw2}

\item{start_year}{first year to estimate}

\item{end_year}{last year to estimate}
}
\description{
Fit log-logistic model
}
