% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_exponential.R
\name{piecewise_exponential}
\alias{piecewise_exponential}
\title{Fit piecewise exponential model}
\usage{
piecewise_exponential(
  data_vr,
  data_direct,
  time_model = "pspline",
  count_model = "poisson",
  start_year = 1950,
  end_year = 2030,
  include_iid_in_pred = F
)
}
\arguments{
\item{data_vr}{input data counts}

\item{data_direct}{input data in pre-processed form}

\item{time_model}{spline or rw2}

\item{count_model}{poisson or poisson_lognormal}

\item{start_year}{first year to estimate}

\item{end_year}{last year to estimate}

\item{include_iid_in_pred}{include IID errors in prediction}
}
\value{
data.frame with predictions
}
\description{
Fit piecewise exponential model
}
